+++ griffon-app/conf/BuildConfig.groovy
@@ -11,7 +11,7 @@ griffon.project.dependency.resolution = {
 //        flatDir name: 'groovyfxPluginLib', dirs: ["${basePath}lib"]
     }
     dependencies {
-        compile 'org.codehaus.groovyfx:groovyfx:0.2-SNAPSHOT'
+        compile 'org.codehaus.groovyfx:groovyfx:0.2'
     }
 }